Extracted popstate/pushState to its own hook
